Consolidate cart actions and reducers
