+++ src/shop-checkout.js
@@ -11,7 +11,7 @@ import { timeOut } from '../node_modules/@polymer/polymer/lib/utils/async.js';
 
 import { store, installReducers } from './shop-redux-store.js';
 import { pushState } from './shop-redux-router.js';
-import { clearCart } from './shop-redux-actions.js';
+import { clearCart } from './shop-redux-cart.js';
 
 installReducers({
   // Internal state from checkout flow (init/success/error).
