+++ src/shop-detail.js
@@ -8,7 +8,7 @@ import { Debouncer } from '../node_modules/@polymer/polymer/lib/utils/debounce.j
 import { microTask } from '../node_modules/@polymer/polymer/lib/utils/async.js';
 
 import { store } from './shop-redux-store.js';
-import { addCartItem } from './shop-redux-actions.js';
+import { addCartItem } from './shop-redux-cart.js';
 import { findCategory, findItem } from './shop-redux-helpers.js';
 
 class ShopDetail extends Element {
